Ensure atom exists before atom from_string_test
`atom_test.from_string_test/0` uses `ok` to test `atom.from_string/0`. However,
the availability of the `ok` atom seems to be a side effect of running the test
trough eunit.

This patch explicitly creates an atom to make sure it exists before using it to
test `atom.from_string/1`. It removes the assertion on `expect`, which had the
same issue and doesn't seem to test another case.