Implement 'gleam/float.round' for JS
We have an issue where JavaScript's Math.round behaves differently to
erlang's round when it comes to negative numbers with 0.5 as the
decimal.

I've added an extra check to make sure that we stay with erlangs
behaviour but it might be considered too expensive and there might be
better ways.